backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.38.0
python-multipart==0.0.20
aiofiles==24.1.0
websockets==12.0
pydantic==2.12.4
pydantic-settings==2.12.0
python-docx==1.1.0
fastmcp==2.12.5
openai==1.51.2
httpx==0.28.1
python-dotenv==1.2.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
